[hashing] Universal Prime-Length Hasher onto F_q
- Created a new `UniversalPrimeLengthHasher` class in the `hashing`
module, which uses the Carter-Wegman universal hash construction to
provide support for hashing an arbitrary-length message provided as a
byte-string onto the prime-order field `F_q` for a prime `q`.
  - Note: This algorithm produces non-deterministic hashes, due to its
  use of random coefficients in: `H_q(x, p) := (ax + b mod p) mod q`, so
  if deterministic, repeatable hashes are needed for the same message
  then a different algorithm should be used.

- Renamed the existing `PrimeLengthTruncatedHasher` class to
`PrimeBasedTruncatedHasher` to more accurately reflect the fact that
this hashing class outputs cryptographic hashes with the same bit-length
as the provided prime `q`, but does not actually map onto the prime-
order field `F_q`.
  - The `schnorr`, `two_party_schnorr` & `nizk_equal_dlogs_proof`
  modules have been updated accordingly, to use the renamed prime based
  bit-length truncated hasher class.
  - Note: This class is still useful for implementing cryptographic
  protocols that rely on this particular hash construction, such as the
  NIST FIPS-186.5 standard for ECDSA signatures.

- Created a new abstract base class (ABC) `ReducedRangeHasher` in the
`hashing` module, which specifies a set of common methods for reduced-
range hashers, and updated both the `PrimeBasedTruncatedHasher` & new
`UniversalPrimeLengthHasher` classes to implement this new interface.

GI-15