[schnorr] Revise membership check re: Schnorr contexts
- Revised the `SchnorrContext` & `TwoPartySchnorrContext` context
classes' methods for converting an ECC point to an `ECC.EccKey` public
key, to use the new elliptic curve point membership check of
`WeierstrassEllipticCurveConfig` for verifying the provided ECC point is
on the configured elliptic curve prior to attempting to construct such a
public key.
  - These context classes now raise a ValueError if the provided ECC
  point does not lie on the configured curve.

- Similarly updated the `TwoPartySchnorrContext` class's
`verify_ecc_point(...)` method to use this new point membership check,
instead of relying on the PyCryptodome library's `Crypto.PublicKey.ECC`
package's `construct(EccPoint)` factory for `EccKey` to perform this
check.

- Note: These changes avoid relying on functionality of the PyCryptodome
library, which is currently used for elliptic curve (ECC) operations,
that should arguably be considered an implementation detail, since this
point membership check is not documented for this `construct(EccPoint)`
factory function or the associated constructor for `EccPoint` that
currently performs this check.

GI-15