[ecc] Construct random generators [wip]
- Added support for constructing random elliptic curve generators, for
which nobody knows the discrete logarithm (w.r.t. a curve's published
generator `G`). When distinct nonce values are used, this algorithm will
produce effectively independent generators (e.g., as required for ECC
Pedersen commitments). [WIP]

- Added a new `number_theory` module, which initially provides a
function that calculates the Legendre symbol `( a / p )`, where `p` is
prime, which returns 1 if `a` is a quadratic residue modulo `p`
(i.e., there exists some `s` such that `s^2 = a mod p` for integer `a`),
or -1 if `a` is a quadratic non-residue modulo `p`, or 0 if `p` divides
`a`.
  - Also, added helper functions `is_quadratic_residue(a, p)` &
`is_quadratic_nonresidue(a, p)`.

- Started implementing a modular square-root function `mod_sqrt(...)`
for this new `number_theory` module, which will use the Tonelli-Shanks
algorithm. [WIP]
  - It initially only supports prime moduli `p` where `p % 4 == 3`, and
will currently raise a `NotImplementedError` if instead `p % 4 == 1`,
since this case requires the implementation of the Tonelli-Shanks
algorithm.
  - (Note: The NIST P-256 elliptic curve's coefficient modulus `n`
satisfies `n % 4 == 3`, so this partly implemented version is still
useful for some testing.)

- Improved the efficiency of the `WeierstrassEllipticCurveConfig`
class's `is_point_on_curve(...)` method, by using the builtin `pow(...)`
function's support for modular exponentiation (that's more efficient
than `(a ** e) % p`).

GI-15