[ecc] Weierstrass curve config. factory by name
- Added support for constructing a `WeierstrassEllipticCurveConfig` ECC
context by elliptic curve common/primary name or an alias, in addition
to the existing support for factory functions dedicated to a specific
curve. Also added a class method for determining if a given curve is
supported, which takes a common/primary name or alias.
  - Note: This new factory function is esp. useful for constructing an
ECC context from a serialized curve name (e.g., when deserializing a
serialized NIZK proof that has been serialized to a string or binary
encoding).

GI-11