Improve mod_sqrt() efficiency
- Improved the efficiency of the `mod_sqrt(...)` function in the
`number_theory` module, by using a special-case identity of Legendre for
efficiently computing the two modular square roots when the prime
modulus `p` is congruent to 5 mod 8.
  - This special-case identity of Legendre requires determining whether
  the integer, for which the modular square root is being calculated, is
  a quartic residue mod `p`.

- Added support for calculating the quartic (or biquadratic) residue
symbol, including adding helper functions for determining if an integer
is a quartic residue or non-residue in re: to a prime modulus.

GI-15