Correct universal prime-length hasher re: XOF case
- Corrected the calculation of the `UniversalPrimeLengthHasher`
universal prime-length hasher's intermediate full-length hash, which had
erroneously initialized the wrong hasher for a provided domain separator
when an XOF was used.

- Corrected the initialization of the NUMS elliptic curve point
generator's context class, which no longer takes a `hash_algorithm`
parameter, as part of initializing a Pedersen commitment's context.

- Revised the Pedersen commitment's context class to use a default nonce
of 0, for the NUMS point generator, instead of 1.

GI-15