[he] Support re-blinding of Paillier ciphertexts
- Added support for re-blinding of Paillier ciphertexts via a new
`obfuscate(...)` method, which provides for indistinguishability of
ciphertexts following one or more homomorphic operations, and is useful
for ensuring privacy when used in secure multi-party computation (MPC)
protocols.

- Revised the homomorphic "scalar" multiplication support to ensure the
scalar multiplier is a non-negative integer, as required by Paillier's
native support for only non-negative integer plaintexts.

- Added `add(...)` and `multiply(...)` methods to corresponding to the
existing dunder methods for '+' and '*' operators.

- Implemented the `__rmul__(...)` dunder method, enabling support for
right-multiply with the '*' operator, which allows the "scalar" integer
multiplier to appear on the left side of the operator.

- Improved doc-comments for homomorphic operations in the Paillier
`EncryptedUnsignedInteger` class.

GI-20