Fix .whl packages re: top-level package discovery
- Corrected the pyproject.toml file, to correctly specify the location
of the top-level package, in the .whl file to be produced, as well as
moving several project-level config. properties to this TOML configs.'
'tool.poetry' section (i.e., instead of the 'project' section), as only
one of these should be present & active where at the sime time.

- Packaging to a .war archive via Poetry is now producing an installable
.whl, where package imports are now working as expected when this .whl
has been installed by another project.