[ecc] Common interface for ECC Points [wip]
- Created an abstract base class for (2D) elliptic curve points,
including support for point addition, negation, scalar multiplication,
and serialization/deserialization, in addition to methods for
determining if a point is the identity element (i.e., the "point-at-
infinity"), as well as class methods for obtaining the base (common
generator) point & this identity point.

- Added an implementation of this `EccPoint2D` ABC that wraps an
`ECC.EccPoint` provided by the PyCryptodome library, and features a
`WeierstrassEllipticCurveConfig` curve configuration.

GI-29