[he] Handle zero scalar special cases re: Paillier
- Revised homomorphic scalar multiplication in the `paillier` encryption
module, to ensure the ciphertext is re-blinded to avoid leaking the
resulting plaintext value 0.

- Also revised the module's homomorphic scalar addition (i.e., scalar
encryption followed by homomorphic addition), to ensure the ciphertext
is re-blinded in the case the scalar summand is 0, to avoid leaking the
fact that the homomorphic sum equals the original ciphertext's plaintext
integer value (i.e., retaining indistinguishability of ciphertexts), due
otherwise to production of identical ciphertexts.

- Updated PyDoc comments for both revised homomorphic operations, to
clarify the special handling for these zero scalar special cases.

GI-20